Refactor to support multiple products within the same state classes
