+++ src/main/scala/com/gu/support/workers/model/Periods.scala
@@ -1,0 +1,9 @@
+package com.gu.support.workers.model
+
+sealed trait Period
+
+case object Monthly extends Period
+
+case object Quarterly extends Period
+
+case object Annual extends Period
