--- src/main/scala/com/gu/support/workers/model/monthlyContributions/state/CreatePaymentMethodState.scala
@@ -1,14 +1,0 @@
-package com.gu.support.workers.model.monthlyContributions.state
-
-import java.util.UUID
-
-import com.gu.support.workers.model.monthlyContributions.Contribution
-import com.gu.support.workers.model._
-
-case class CreatePaymentMethodState(
-  requestId: UUID,
-  user: User,
-  contribution: Contribution,
-  paymentFields: PaymentFields,
-  acquisitionData: Option[AcquisitionData]
-) extends StepFunctionUserState
