--- src/main/scala/com/gu/support/workers/model/monthlyContributions/state/CreateSalesforceContactState.scala
@@ -1,14 +1,0 @@
-package com.gu.support.workers.model.monthlyContributions.state
-
-import java.util.UUID
-
-import com.gu.support.workers.model.monthlyContributions.Contribution
-import com.gu.support.workers.model.{AcquisitionData, PaymentMethod, User}
-
-case class CreateSalesforceContactState(
-  requestId: UUID,
-  user: User,
-  contribution: Contribution,
-  paymentMethod: PaymentMethod,
-  acquisitionData: Option[AcquisitionData]
-) extends StepFunctionUserState
