+++ src/main/scala/com/gu/support/workers/model/states/CompletedState.scala
@@ -1,9 +1,8 @@
-package com.gu.support.workers.model.monthlyContributions.state
+package com.gu.support.workers.model.states
 
 import java.util.UUID
 
-import com.gu.support.workers.model._
-import com.gu.support.workers.model.monthlyContributions.{Contribution, Status}
+import com.gu.support.workers.model.{Status, _}
 
 case class CompletedState(
   requestId: UUID,
