+++ src/main/scala/com/gu/support/workers/model/states/CreateZuoraSubscriptionState.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model.monthlyContributions.state
+package com.gu.support.workers.model.states
 
 import java.util.UUID
 
@@ -8,7 +8,7 @@ import com.gu.support.workers.model.{AcquisitionData, PaymentMethod, SalesforceC
 case class CreateZuoraSubscriptionState(
   requestId: UUID,
   user: User,
-  contribution: Contribution,
+  product: Product,
   paymentMethod: PaymentMethod,
   salesForceContact: SalesforceContactRecord,
   acquisitionData: Option[AcquisitionData]
