+++ src/main/scala/com/gu/support/workers/model/states/StepFunctionUserState.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model.monthlyContributions.state
+package com.gu.support.workers.model.states
 
 import com.gu.support.workers.model.User
